osutil: split group lookup into cgo and !cgo stubs
Querying UNIX groups relies of C library calls getgrnam_r and getgrgid_r.
The code using osutil group API will not build with CGO_ENABLED=0. Provide stubs
that implement low-level calls and return known errors in such case.
