+++ osutil/group_stubs_test.go
@@ -1,0 +1,42 @@
+// -*- Mode: Go; indent-tabs-mode: t -*-
+
+// +build !cgo
+
+/*
+ * Copyright (C) 2017 Canonical Ltd
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package osutil
+
+import (
+	. "gopkg.in/check.v1"
+)
+
+type groupStubsSuite struct {
+}
+
+var _ = Suite(&groupStubsSuite{})
+
+func (s *groupStubsSuite) TestFindGroupStubs(c *C) {
+	_, err := FindGroup(0)
+	c.Check(err, ErrorMatches, "lookupGroupByGid not implemented for non cgo builds")
+
+	_, err = FindGid("root")
+	c.Check(err, ErrorMatches, "lookupGroup not implemented for non cgo builds")
+
+	_, err = FindGroupOwning("/etc/passwd")
+	c.Check(err, ErrorMatches, "lookupGroupByGid not implemented for non cgo builds")
+}
